casted the time internal in orientation delay to the appropriate format.

+++ SwrveConversationSDK/Conversation/SwrveConversationItemViewController.m
@@ -428,7 +428,7 @@
         }
         currentOrientation = orientation;
         //delay for .01ms to account for rotation frame returned being out of date
-        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 0.01 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
+        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (u_int64_t)0.01 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
             // Tell everyone who needs to know that orientation has changed, individual items will react to this and change shape
             for(SwrveConversationAtom *atom in self.conversationPane.content) {
                 if([atom.delegate respondsToSelector:@selector(respondToDeviceOrientationChange:)]) {
